+++ projects/ngx-location-picker/src/lib/services/ngx-location-picker.helper.spec.ts
@@ -20,33 +20,50 @@ describe('NgxLocationPickerHelper', () => {
         const testCases = [
             {
                 query: '',
-                expectedResult: false
+                expectedResult: false,
+                keywords: []
             },
             {
                 query: 'Groenplaats',
-                expectedResult: false
+                expectedResult: false,
+                keywords: ['kaainummer']
             },
             {
                 query: 'Generaal armstrongweg',
-                expectedResult: false
+                expectedResult: false,
+                keywords: ['kaainummer']
+
             },
             {
                 query: 'Generaal armstrongweg 1',
-                expectedResult: true
+                expectedResult: true,
+                keywords: ['kaainummer']
             },
             {
                 query: 'Generaal armstrongweg 1A',
-                expectedResult: true
+                expectedResult: true,
+                keywords: ['kaainummer']
             },
             {
                 query: 'Kaainummer 757',
-                expectedResult: false
+                expectedResult: false,
+                keywords: ['kaainummer']
+            },
+            {
+                query: 'Kaainummer 15',
+                expectedResult: false,
+                keywords: ['Kaainummer', 'straatnaam']
+            },
+            {
+                query: 'Kaainummer 20',
+                expectedResult: true,
+                keywords: []
             }
         ];
 
         testCases.forEach((test, index) => {
-            it(`should test if '${test.query}' isAddress with expectedResult: ${test.expectedResult}`, () => {
-                const result = service.isAddress(test.query);
+            it(`should test if '${test.query}' isAddress with expectedResult: ${test.expectedResult} and keywords: ${test.keywords}`, () => {
+                const result = service.isAddress(test.query, test.keywords);
                 expect(result).toEqual(test.expectedResult);
             });
         })