+++ projects/ngx-location-picker/src/lib/services/ngx-location-picker.helper.ts
@@ -2,7 +2,6 @@ import {Injectable} from '@angular/core';
 import {HttpParams} from '@angular/common/http';
 import {LambertModel, LocationModel} from '../types/location.model';
 import {AddressQueryModel} from '../types/address-query.model';
-import {CascadingCoordinateRulesModel, CascadingCoordinateRulesType} from '../types/cascading-rules.model';
 import proj4 from 'proj4';
 import { AddressModel } from '../types/address.model';
 import { CoordinateModel } from '../types/coordinate.model';
@@ -28,11 +27,6 @@ export class NgxLocationPickerHelper {
   maxX = 51.6;
   minY = 2.3;
   maxY = 6.45;
-
-  // If search string contains one of these words, search for locations instead of address
-  private locationKeywords: string [] = [
-    'kaainummer'
-  ]
 
   /**
    * Converts a query object to HttpParams.
@@ -106,13 +100,15 @@ export class NgxLocationPickerHelper {
    *
    * @return boolean
    */
-  isAddress(query: string): boolean {
+  isAddress(query: string, locationKeywords: string[]): boolean {
     const addressParts: Array<string> | null = (query) ? query.split(' ') : null;
+
+    const lowerLocationKeyWords = locationKeywords.map(x => x.toLowerCase());
 
     if (addressParts && Array.isArray(addressParts) && addressParts.length > 1) {
       for (const [index, value] of addressParts.entries()) {
         // exclusion for specific search combinations (ex 'kaainummer + number' GIS-537)
-        if (this.locationKeywords.includes(addressParts[index].toLowerCase()))
+        if (lowerLocationKeyWords.includes(addressParts[index].toLowerCase()))
         {
           return false;
         }