+++ projects/ngx-location-picker/src/lib/types/delegate-search.model.ts
@@ -14,4 +14,5 @@ export interface DelegateSearchModel {
     cascadingCoordinateLimit: number;
     cascadingCoordinateRules: CascadingCoordinateRulesModel[];
     selectedLocation: LocationModel | AddressModel | CoordinateModel;
+    locationKeywords: string[];
 }